Encapsulate signalR hub connection instance
