+++ webapp/src/redux/app/store.ts
@@ -3,11 +3,7 @@
 import { AnyAction, Dispatch, MiddlewareAPI, MiddlewareArray, ThunkMiddleware, configureStore } from '@reduxjs/toolkit';
 import { AppState } from '../features/app/AppState';
 import { ConversationsState } from '../features/conversations/ConversationsState';
-import {
-    registerSignalREvents,
-    signalRMiddleware,
-    startSignalRConnection,
-} from '../features/message-relay/signalRMiddleware';
+import { signalRMiddleware } from '../features/message-relay/signalRMiddleware';
 import { PluginsState } from '../features/plugins/PluginsState';
 import { UsersState } from '../features/users/UsersState';
 import resetStateReducer, { resetApp } from './rootReducer';
@@ -35,11 +31,6 @@ export interface RootState {
 export const getSelectedChatID = (): string => {
     return store.getState().conversations.selectedId;
 };
-
-// Start the signalR connection to make sure messages are
-// sent to all clients and received by all clients
-startSignalRConnection(store);
-registerSignalREvents(store);
 
 export type AppDispatch = typeof store.dispatch;
